inital commit

+++ core/src/com/brett/game/states/PlayState.java
@@ -30,7 +30,8 @@ public class PlayState extends State {
         groundPos1 = new Vector2(cam.position.x - cam.viewportWidth / 2, GROUND_Y_OFFSET);
         groundPos2 = new Vector2((cam.position.x - cam.viewportWidth / 2) + ground.getWidth(), GROUND_Y_OFFSET);
 
-        cam.setToOrtho(false, FlappyDemo.WIDTH / 2, FlappyDemo.HEIGHT / 2);
+        cam.setToOrtho(false, (float)FlappyDemo.WIDTH / 2, (float)FlappyDemo.HEIGHT / 2);
+
 
         tubes = new Array<Tube>();
